# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "booking list",
                "operationId": "showBookingList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Add Booking",
                "description": "Booking Add Here",
                "operationId": "addBooking",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "time_slot_id"
                                ],
                                "properties": {
                                    "time_slot_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Escape room detail",
                "operationId": "showBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/escape-rooms": {
            "get": {
                "tags": [
                    "EscapeRoom"
                ],
                "summary": "escape room list",
                "operationId": "showRoomList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/escape-rooms/{id}": {
            "get": {
                "tags": [
                    "EscapeRoom"
                ],
                "summary": "escape room detail",
                "operationId": "showRoomDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "room id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/escape-rooms/time-slots/{id}": {
            "get": {
                "tags": [
                    "EscapeRoom"
                ],
                "summary": "escape room available time slots",
                "operationId": "showRoomTimeSlots",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "room id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}